Refactored out profiling code into separate file
