+++ vpr/SRC/route/route_profiling.h
@@ -1,0 +1,29 @@
+#pragma once
+
+namespace profiling {
+
+// action counters for what actions setting up routing resources took to build targets list
+void net_rerouted();
+void route_tree_pruned();
+void route_tree_preserved();
+
+
+// timing functions where *_start starts a clock and *_end terminates the clock
+void sink_criticality_start();
+void sink_criticality_end(float target_criticality);
+
+void net_rebuild_start();
+void net_rebuild_end(unsigned net_fanout, unsigned sinks_left_to_route);
+
+void net_fanout_start();
+void net_fanout_end(unsigned net_fanout);
+
+
+// analysis functions for printing out the profiling data for an iteration
+void congestion_analysis();
+void time_on_criticality_analysis();
+void time_on_fanout_analysis();
+
+void profiling_initialization(unsigned max_net_fanout);
+
+} // end namespace profiling
